fix: Moved 2 classes from "model" to "modelDice"
